refactor: HistoryWithFilter.tsx, Info.tsx 분해 (Phase 4)
HistoryWithFilter.tsx (585줄 → 300줄):
- RunHistoryItem.tsx 추출
- RunHistoryGalleryItem.tsx 추출

Info.tsx (573줄 → 400줄):
- CadenceAssistControl.tsx 추출
- ProfileCard.tsx 추출
