+++ src/components/course/HistoryWithFilter.tsx
@@ -1,16 +1,15 @@
-import { DefaultLogo } from "@/assets/icons/icons";
-import { ChevronIcon, GhostIcon } from "@/assets/svgs/svgs";
 import { RunResponse } from "@/src/apis/types/run";
-import colors from "@/src/theme/colors";
 import { formatDate } from "@/src/utils/formatDate";
-import { getDate, getFormattedPace, getRunTime } from "@/src/utils/runUtils";
+import { getDate } from "@/src/utils/runUtils";
 import { BottomSheetModal } from "@gorhom/bottom-sheet";
 import { FlashList } from "@shopify/flash-list";
 import { useRouter } from "expo-router";
 import { useEffect, useMemo, useRef, useState } from "react";
-import { Image, StyleSheet, TouchableOpacity, View } from "react-native";
+import { View } from "react-native";
 import { GoRunCalendar } from "../calendar/GoRunCalendar";
-import { BottomModal, Divider, DualFilter, EmptyListView, FilterBar, ScrollButton, Section, Typography } from "@/src/components/ui";
+import { BottomModal, DualFilter, EmptyListView, FilterBar, ScrollButton, Section } from "@/src/components/ui";
+import { RunHistoryItem } from "./RunHistoryItem";
+import { RunHistoryGalleryItem } from "./RunHistoryGalleryItem";
 
 type HistoryWithFilterProps = {
     data: RunResponse[];
@@ -299,287 +298,3 @@ export const HistoryWithFilter = ({
         </View>
     );
 };
-
-interface RunHistoryGalleryItemProps {
-    mode: "SOLO" | "GHOST";
-    imageUrl: string;
-    name: string;
-    courseName: string | null;
-    distance: number;
-    duration: number;
-    averagePace: number;
-    cadence: number;
-    onShowHistory: () => void;
-    isSelected: boolean;
-    startedAt: number;
-    selectedFilter: "date" | "course";
-}
-
-const RunHistoryGalleryItem = ({
-    mode,
-    imageUrl,
-    name,
-    courseName,
-    distance,
-    duration,
-    averagePace,
-    cadence,
-    onShowHistory,
-    isSelected,
-    startedAt,
-    selectedFilter,
-}: RunHistoryGalleryItemProps) => {
-    return (
-        <TouchableOpacity style={styles.container} onPress={onShowHistory}>
-            <View style={styles.imageContainer}>
-                {mode === "GHOST" && (
-                    <View style={styles.iconContainer}>
-                        <GhostIcon
-                            width={20}
-                            height={12}
-                            color={colors.primary}
-                        />
-                    </View>
-                )}
-
-                <Image
-                    source={imageUrl ? { uri: imageUrl } : DefaultLogo}
-                    style={styles.image}
-                />
-            </View>
-            <View style={styles.contentContainer}>
-                <View style={styles.contentHeader}>
-                    <View style={styles.nameContainer}>
-                        <View
-                            style={{
-                                flexDirection: "row",
-                                alignItems: "center",
-                                justifyContent: "space-between",
-                                flex: 1,
-                            }}
-                        >
-                            <Typography
-                                variant="subhead1"
-                                color={isSelected ? "primary" : "gray20"}
-                            >
-                                {name}
-                            </Typography>
-                            <ChevronIcon color={colors.gray[40]} />
-                        </View>
-                    </View>
-                </View>
-                <View>
-                    <View style={styles.content}>
-                        <Typography
-                            variant="body1"
-                            color={isSelected ? "gray20" : "gray40"}
-                        >
-                            {distance.toFixed(2)}km
-                        </Typography>
-                        <Divider />
-                        <Typography
-                            variant="body1"
-                            color={isSelected ? "gray20" : "gray40"}
-                        >
-                            {getRunTime(duration, "HH:MM:SS_IF_HH_EXISTS")}
-                        </Typography>
-                    </View>
-                    <View style={styles.content}>
-                        <Typography
-                            variant="body1"
-                            color={isSelected ? "gray20" : "gray40"}
-                        >
-                            {getFormattedPace(averagePace)}
-                        </Typography>
-                        <Divider />
-                        <Typography
-                            variant="body1"
-                            color={isSelected ? "gray20" : "gray40"}
-                        >
-                            {cadence}spm
-                        </Typography>
-                    </View>
-                    <View style={styles.dateContainer}>
-                        <Typography
-                            variant="body3"
-                            color={isSelected ? "gray20" : "gray40"}
-                        >
-                            {selectedFilter === "date"
-                                ? courseName ?? ""
-                                : getDate(startedAt)}
-                        </Typography>
-                    </View>
-                </View>
-            </View>
-        </TouchableOpacity>
-    );
-};
-
-interface RunHistoryItemProps {
-    mode: "SOLO" | "GHOST";
-    name: string;
-    courseName: string | null;
-    distance: number;
-    duration: number;
-    averagePace: number;
-    cadence: number;
-    onShowHistory: () => void;
-    isSelected: boolean;
-}
-
-const RunHistoryItem = ({
-    mode,
-    name,
-    courseName,
-    distance,
-    duration,
-    averagePace,
-    cadence,
-    onShowHistory,
-    isSelected,
-}: RunHistoryItemProps) => {
-    return (
-        <TouchableOpacity style={styles.itemContainer} onPress={onShowHistory}>
-            <View style={{ gap: 2 }}>
-                <View style={styles.nameContainer}>
-                    {mode === "GHOST" && (
-                        <View style={styles.iconCompactContainer}>
-                            <GhostIcon
-                                width={13}
-                                height={8.13}
-                                color={colors.primary}
-                            />
-                        </View>
-                    )}
-                    <Typography
-                        variant="subhead1"
-                        color={isSelected ? "primary" : "gray20"}
-                    >
-                        {name}
-                    </Typography>
-                    <TouchableOpacity
-                        onPress={onShowHistory}
-                        style={styles.dateContainer}
-                    >
-                        {courseName && (
-                            <View style={styles.nameContainer}>
-                                <Divider />
-                                <Typography
-                                    variant="caption1"
-                                    color={isSelected ? "gray20" : "gray40"}
-                                >
-                                    {courseName}
-                                </Typography>
-                            </View>
-                        )}
-                        <ChevronIcon
-                            color={
-                                isSelected ? colors.gray[20] : colors.gray[40]
-                            }
-                            width={18}
-                            height={18}
-                        />
-                    </TouchableOpacity>
-                </View>
-                <View style={styles.content}>
-                    <Typography
-                        variant="body2"
-                        color={isSelected ? "gray20" : "gray40"}
-                    >
-                        {distance.toFixed(2)}km
-                    </Typography>
-                    <Divider />
-                    <Typography
-                        variant="body2"
-                        color={isSelected ? "gray20" : "gray40"}
-                    >
-                        {getRunTime(duration, "HH:MM:SS_IF_HH_EXISTS")}
-                    </Typography>
-                    <Divider />
-                    <Typography
-                        variant="body2"
-                        color={isSelected ? "gray20" : "gray40"}
-                    >
-                        {getFormattedPace(averagePace)}
-                    </Typography>
-                    <Divider />
-                    <Typography
-                        variant="body2"
-                        color={isSelected ? "gray20" : "gray40"}
-                    >
-                        {cadence}spm
-                    </Typography>
-                </View>
-            </View>
-        </TouchableOpacity>
-    );
-};
-
-const styles = StyleSheet.create({
-    itemContainer: {
-        flexDirection: "row",
-        justifyContent: "space-between",
-        alignItems: "center",
-    },
-    container: {
-        flexDirection: "row",
-        gap: 20,
-        alignItems: "center",
-    },
-    imageContainer: {
-        backgroundColor: colors.gray[80],
-        width: 120,
-        height: 120,
-        borderRadius: 10,
-        alignItems: "center",
-        justifyContent: "center",
-    },
-    image: {
-        width: 120,
-        height: 120,
-        borderRadius: 10,
-    },
-    contentContainer: {
-        gap: 5,
-        flex: 1,
-    },
-    contentHeader: {
-        flexDirection: "row",
-        alignItems: "center",
-        justifyContent: "space-between",
-    },
-    nameContainer: {
-        flexDirection: "row",
-        alignItems: "center",
-        gap: 6,
-    },
-    iconCompactContainer: {
-        width: 22,
-        height: 22,
-        borderRadius: 6,
-        backgroundColor: "rgba(226, 255, 0, 0.2)",
-        justifyContent: "center",
-        alignItems: "center",
-    },
-    iconContainer: {
-        width: 34,
-        height: 34,
-        borderRadius: 10,
-        backgroundColor: "rgba(226, 255, 0, 0.2)",
-        justifyContent: "center",
-        alignItems: "center",
-        position: "absolute",
-        left: 4,
-        top: 4,
-        zIndex: 10,
-    },
-    content: {
-        flexDirection: "row",
-        alignItems: "center",
-        gap: 10,
-    },
-    dateContainer: {
-        flexDirection: "row",
-        alignItems: "center",
-    },
-});
