+++ src/components/profile/Info.tsx
@@ -1,4 +1,3 @@
-import { DefaultProfileIcon } from "@/assets/icons/icons";
 import { ChevronIcon } from "@/assets/svgs/svgs";
 import {
     getPresignedUrl,
@@ -22,18 +21,18 @@ import { useFocusEffect, useRouter } from "expo-router";
 import { useCallback, useState } from "react";
 import {
     Alert,
-    Image,
     Linking,
     RefreshControl,
     ScrollView,
-    StyleSheet,
     TouchableOpacity,
     View,
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { ProfileNoticeSection } from "../notice/ui/ProfileNoticeSection";
 import { CadenceAssistGuide } from "../onboarding/CadenceAssistGuide";
-import { Divider, ListSectionContainer, ListSectionItem, StyledButton, StyledSwitch, Typography, showToast } from "@/src/components/ui";
+import { ListSectionContainer, ListSectionItem, StyledButton, StyledSwitch, Typography, showToast } from "@/src/components/ui";
+import { CadenceAssistControl } from "./CadenceAssistControl";
+import { ProfileCard } from "./ProfileCard";
 
 export const Info = ({
     setModalType,
@@ -246,7 +245,7 @@ export const Info = ({
             >
                 {/* Profile */}
                 <View style={{ gap: 15, marginTop: 10 }}>
-                    <Profile userInfo={userInfo ?? null} loading={isFetching} />
+                    <ProfileCard userInfo={userInfo ?? null} loading={isFetching} />
                     <View style={{ flexDirection: "row", gap: 4 }}>
                         <StyledButton
                             title="프로필 이미지 변경"
@@ -399,175 +398,3 @@ export const Info = ({
         </>
     );
 };
-
-const CadenceAssistControl = ({ isEnabled }: { isEnabled: boolean }) => {
-    const {
-        cadenceTarget: value,
-        decCadenceTarget,
-        incCadenceTarget,
-    } = useLocalPrefs();
-    return (
-        <View style={styles.cadenceAssistControl}>
-            {/* -10 버튼 */}
-            <TouchableOpacity
-                disabled={!isEnabled}
-                onPress={() => decCadenceTarget(10)}
-                onLongPress={() => decCadenceTarget(10)}
-                style={[
-                    styles.cadenceAssistButton,
-                    isEnabled ? {} : styles.disabledCadenceAssistControl,
-                ]}
-            >
-                <Typography
-                    variant="subhead3"
-                    color={isEnabled ? "white" : "gray60"}
-                >
-                    -10
-                </Typography>
-            </TouchableOpacity>
-
-            {/* 현재 값 */}
-            <View
-                style={[
-                    styles.cadenceAssistPanel,
-                    isEnabled ? {} : styles.disabledCadenceAssistControl,
-                ]}
-            >
-                <Typography
-                    variant="subhead3"
-                    color={isEnabled ? "white" : "gray60"}
-                >
-                    {value} spm
-                </Typography>
-            </View>
-
-            {/* +10 버튼 */}
-            <TouchableOpacity
-                disabled={!isEnabled}
-                onPress={() => incCadenceTarget(10)}
-                onLongPress={() => incCadenceTarget(10)}
-                style={[
-                    styles.cadenceAssistButton,
-                    isEnabled ? {} : styles.disabledCadenceAssistControl,
-                ]}
-            >
-                <Typography
-                    variant="subhead3"
-                    color={isEnabled ? "white" : "gray60"}
-                >
-                    +10
-                </Typography>
-            </TouchableOpacity>
-        </View>
-    );
-};
-
-const Profile = ({
-    userInfo,
-    loading,
-}: {
-    userInfo: GetUserInfoResponse | null;
-    loading?: boolean;
-}) => {
-    const userProfileImageUrl =
-        userInfo?.profilePictureUrl?.split("?X-Amz-")[0];
-
-    return (
-        <View style={styles.profileContent}>
-            <Image
-                source={
-                    userProfileImageUrl
-                        ? { uri: userProfileImageUrl }
-                        : DefaultProfileIcon
-                }
-                style={styles.profileImage}
-            />
-            <View>
-                <Typography variant="headline" color="gray20">
-                    {loading ? "" : userInfo?.nickname ?? "고스트러너"}
-                </Typography>
-                <View style={styles.profileInfo}>
-                    <Typography variant="body2" color="gray40">
-                        {loading
-                            ? ""
-                            : userInfo?.height
-                            ? `${userInfo.height}cm`
-                            : "키 비공개"}
-                    </Typography>
-
-                    <Divider />
-                    <Typography variant="body2" color="gray40">
-                        {loading
-                            ? ""
-                            : userInfo?.weight
-                            ? `${userInfo.weight}kg`
-                            : "몸무게 비공개"}
-                    </Typography>
-
-                    <Divider />
-                    <Typography variant="body2" color="gray40">
-                        {loading
-                            ? ""
-                            : userInfo?.gender === "MALE"
-                            ? "남성"
-                            : userInfo?.gender === "FEMALE"
-                            ? "여성"
-                            : "성별 비공개"}
-                    </Typography>
-                </View>
-            </View>
-        </View>
-    );
-};
-
-const styles = StyleSheet.create({
-    cadenceAssistControl: {
-        marginTop: -2,
-        paddingHorizontal: 17,
-        paddingBottom: 17,
-        flexDirection: "row",
-        alignItems: "center",
-        gap: 4,
-    },
-    cadenceAssistButton: {
-        height: 32,
-        paddingHorizontal: 12,
-        borderRadius: 6,
-        backgroundColor: colors.gray[80],
-        justifyContent: "center",
-        alignItems: "center",
-        borderWidth: 1,
-        borderColor: "#171717",
-        boxShadow: "0px 2px 6px 0px rgba(0, 0, 0, 0.15)",
-    },
-    cadenceAssistPanel: {
-        flex: 1,
-        height: 32,
-        borderRadius: 6,
-        backgroundColor: "#171717",
-        justifyContent: "center",
-        alignItems: "center",
-        borderWidth: 1,
-        borderColor: colors.gray[80],
-    },
-    disabledCadenceAssistControl: {
-        backgroundColor: "#171717",
-        borderColor: "#171717",
-    },
-
-    profileContent: {
-        flexDirection: "row",
-        gap: 15,
-        alignItems: "center",
-    },
-    profileImage: {
-        width: 60,
-        height: 60,
-        borderRadius: 100,
-    },
-    profileInfo: {
-        flexDirection: "row",
-        gap: 10,
-        alignItems: "center",
-    },
-});
